// mobile/src/api/types.ts
export interface components {
    schemas: {
// ...
        MemberResponse: {
            id: string;
            gymId: string;
            memberCode: string;
            name: string;
            phone: string;
            gender: string;
            /** Format: date */
            dob: string;
            address: string;
            photoUrl: string;
            status: string;
            /** Format: date */
            currentExpiry: string;
            /** Format: int64 */
            currentDuePaise: number;
            /** Format: date */
            lastAttendanceDate: string;
            currentMembership: components["schemas"]["MembershipResponse"];
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        MembershipResponse: {
            id: string;
            gymId: string;
            memberId: string;
            planId: string;
            planName: string;
            /** Format: int64 */
            totalPaise: number;
            /** Format: int64 */
            paidPaise: number;
            /** Format: int64 */
            duePaise: number;
            /** Format: date */
            startDate: string;
            /** Format: date */
            expiryDate: string;
            status: string;
            /** Format: date-time */
            createdAt: string;
        };
// ...
        PagedResponseMemberResponse: {
            content: components["schemas"]["MemberResponse"][];
            /** Format: int32 */
            page: number;
            /** Format: int32 */
            size: number;
            /** Format: int64 */
            totalElements: number;
            /** Format: int32 */
            totalPages: number;
            last: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}